Localization - Fixed dynamic shop link in cart - Added some Spring localization - Refactored some part - Reduced layout shifting during form validation

+++ back-end/src/main/java/com/ring/service/impl/PublisherServiceImpl.java
@@ -1,5 +1,6 @@
 package com.ring.service.impl;
 
+import com.ring.common.AppConstants;
 import com.ring.dto.request.PublisherRequest;
 import com.ring.dto.response.PagingResponse;
 import com.ring.dto.response.publishers.PublisherDTO;
@@ -10,11 +11,12 @@ import com.ring.model.entity.Publisher;
 import com.ring.repository.PublisherRepository;
 import com.ring.service.ImageService;
 import com.ring.service.PublisherService;
-import com.ring.utils.FileUploadUtil;
+import com.ring.common.FileUploadUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.cache.annotation.Caching;
+import org.springframework.context.support.DefaultMessageSourceResolvable;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -31,15 +33,19 @@ public class PublisherServiceImpl implements PublisherService {
 
     private final PublisherRepository pubRepo;
     private final PublisherMapper pubMapper;
-    private final ImageService imageService;
 
-    @Cacheable(cacheNames = "publishers")
+    private final ImageService imageService;
+    private final MessageService messageService;
+
+    @Cacheable(cacheNames = AppConstants.PUBLISHERS)
     public PagingResponse<PublisherDTO> getPublishers(Integer pageNo,
             Integer pageSize,
             String sortBy,
             String sortDir) {
         Pageable pageable = PageRequest.of(pageNo, pageSize,
-                sortDir.equals("asc") ? Sort.by(sortBy).ascending() : Sort.by(sortBy).descending());
+                sortDir.equals(AppConstants.ASCENDING) 
+                    ? Sort.by(sortBy).ascending() 
+                    : Sort.by(sortBy).descending());
 
         // Fetch from database
         Page<Publisher> pubsList = pubRepo.findPublishers(pageable);
@@ -53,11 +59,15 @@ public class PublisherServiceImpl implements PublisherService {
                 pubsList.isEmpty());
     }
 
-    @Cacheable(cacheNames = "publishers")
+    @Cacheable(cacheNames = AppConstants.PUBLISHERS)
     public PagingResponse<PublisherDTO> getRelevantPublishers(Integer pageNo,
             Integer pageSize,
             Integer cateId) {
-        Pageable pageable = PageRequest.of(pageNo, pageSize, Sort.by("id").descending());
+
+        Pageable pageable = PageRequest.of(pageNo, 
+                pageSize, 
+                Sort.by(AppConstants.ID)
+                .descending());
 
         // Fetch from database
         Page<Publisher> pubsList = pubRepo.findRelevantPublishers(cateId, pageable);
@@ -71,25 +81,28 @@ public class PublisherServiceImpl implements PublisherService {
                 pubsList.isEmpty());
     }
 
-    @Cacheable(cacheNames = "publisher", key = "#id")
+    @Cacheable(cacheNames = AppConstants.PUBLISHER, key = "#id")
     public PublisherDTO getPublisher(Integer id) {
+
         Publisher publisher = pubRepo.findWithImageById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Publisher not found!",
-                        "Không tìm thấy nhà xuất bản yêu cầu!"));
+                .orElseThrow(() -> {
+                    var errorMsg = messageService.getMessage("exception.not.found",
+                            new Object[]{ new DefaultMessageSourceResolvable("label.pub") });
+                    return new ResourceNotFoundException(errorMsg);
+                });
 
         PublisherDTO publisherDTO = pubMapper.apply(publisher); // Map to DTO
         return publisherDTO;
     }
 
-    @CacheEvict(cacheNames = "publishers")
+    @CacheEvict(cacheNames = AppConstants.PUBLISHERS)
     @Transactional
-    public Publisher addPublisher(PublisherRequest request,
-            MultipartFile file) {
+    public Publisher addPublisher(PublisherRequest request, MultipartFile file) {
+
         Image image = null;
 
         // Image upload
-        if (file != null)
-            image = imageService.upload(file, FileUploadUtil.ASSET_FOLDER);
+        if (file != null) image = imageService.upload(file, FileUploadUtil.ASSET_FOLDER);
 
         // Create new publisher
         var publisher = Publisher.builder()
@@ -100,22 +113,27 @@ public class PublisherServiceImpl implements PublisherService {
         return addedPub;
     }
 
-    @Caching(evict = { @CacheEvict(cacheNames = "publishers"),
-            @CacheEvict(cacheNames = "publisher", key = "#id") })
+    @Caching(evict = { @CacheEvict(cacheNames = AppConstants.PUBLISHERS),
+            @CacheEvict(cacheNames = AppConstants.PUBLISHER, key = "#id") })
     @Transactional
     public Publisher updatePublisher(Integer id,
             PublisherRequest request,
             MultipartFile file) {
+
         // Get original publisher
         Publisher publisher = pubRepo.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Publisher not found",
-                        "Không tìm thấy nhà xuất bản yêu cầu!"));
+                .orElseThrow(() -> {
+                    var errorMsg = messageService.getMessage("exception.not.found",
+                            new Object[]{ new DefaultMessageSourceResolvable("label.pub") });
+                    return new ResourceNotFoundException(errorMsg);
+                });
 
         // Image upload/replace
         if (file != null) { // Contain new image >> upload/replace
+
             Long imageId = publisher.getImage() != null ? publisher.getImage().getId() : null;
-            if (imageId != null)
-                imageService.deleteImage(imageId); // Delete old image
+
+            if (imageId != null) imageService.deleteImage(imageId); // Delete old image
 
             Image savedImage = imageService.upload(file, FileUploadUtil.ASSET_FOLDER); // Upload new image
             publisher.setImage(savedImage); // Set new image
@@ -128,29 +146,33 @@ public class PublisherServiceImpl implements PublisherService {
         return updatedPub;
     }
 
-    @Caching(evict = { @CacheEvict(cacheNames = "publishers"),
-            @CacheEvict(cacheNames = "publisher", key = "#id") })
+    @Caching(evict = { @CacheEvict(cacheNames = AppConstants.PUBLISHERS),
+            @CacheEvict(cacheNames = AppConstants.PUBLISHER, key = "#id") })
     @Transactional
     public void deletePublisher(Integer id) {
+        
         pubRepo.deleteById(id);
     }
 
-    @CacheEvict(cacheNames = "publishers")
+    @CacheEvict(cacheNames = AppConstants.PUBLISHERS)
     @Transactional
     public void deletePublishers(List<Integer> ids) {
+
         pubRepo.deleteAllByIdInBatch(ids);
     }
 
-    @CacheEvict(cacheNames = "publishers")
+    @CacheEvict(cacheNames = AppConstants.PUBLISHERS)
     @Transactional
     public void deletePublishersInverse(List<Integer> ids) {
+
         List<Integer> listDelete = pubRepo.findInverseIds(ids);
         pubRepo.deleteAllByIdInBatch(listDelete);
     }
 
-    @CacheEvict(cacheNames = "publishers")
+    @CacheEvict(cacheNames = AppConstants.PUBLISHERS)
     @Transactional
     public void deleteAllPublishers() {
+
         pubRepo.deleteAll();
     }
 }
